Fix misunderstanding about passing props into reactstrap input component
The autosuggest component passes props into an input component, if
renderInputComponent prop is used.
These props are different from props passed into inputProps.
This difference was not clear from the code, due to a misstake - the
inputProps were passed into a helper renderInputComponent method, but were
in fact not used.
Therefore the previous first argument 'inputProps' was removed from the
renderInputComponent method.