Specify type for InputForm's custom props
The InputForm component is wrapped by the ReactRouter. Because of that
the ReactRouter is using its own props. Into these props an additional
prop is assigned - called customProps into which component's own props
are assigned.
Because of that the InputForm component has to use an intersection type
as a prop type.